requirements.txt
Flask==3.0.0
Flask_SocketIO==5.3.6
GPUtil==1.4.0
GPUtil==1.4.0
python-socketio==5.10.0
